[Sonic Frontiers] Fix cast in UnmanagedString

+++ RFL2HMM/Templates/Frontiers/BossGiantConfig.cs
@@ -30,7 +30,7 @@ public class BossGiantConfigClass
                 return Marshal.PtrToStringAnsi((nint)pValue);
             }
 
-            set => pValue = Marshal.StringToHGlobalAnsi(value);
+            set => pValue = (long)Marshal.StringToHGlobalAnsi(value);
         }
 
         public UnmanagedString(string in_value)
