+++ RFL2HMM/Templates/Frontiers/SonicParameters.cs
@@ -57,7 +57,7 @@ public class SonicParametersClass
                 return Marshal.PtrToStringAnsi((nint)pValue);
             }
 
-            set => pValue = Marshal.StringToHGlobalAnsi(value);
+            set => pValue = (long)Marshal.StringToHGlobalAnsi(value);
         }
 
         public UnmanagedString(string in_value)
